Refactor DeploymentRecord to ClassRuntime
- Updated the codebase to replace all instances of `DeploymentRecord` with `ClassRuntime` to better reflect the functionality of the resource.
- Modified API endpoints, handlers, and client methods to accommodate the new naming convention.
- Adjusted the README and documentation to reflect changes in resource terminology.
- Updated tests to ensure compatibility with the new ClassRuntime structure.
- Introduced a new Custom Resource Definition (CRD) for ClassRuntime in Kubernetes.
- Enhanced CLI commands to support ClassRuntime operations.
