+++ control-plane/oprc-crm/src/bin/crdgen.rs
@@ -1,8 +1,8 @@
 use kube::core::CustomResourceExt;
-use oprc_crm::crd::deployment_record::DeploymentRecord;
+use oprc_crm::crd::class_runtime::ClassRuntime;
 
 fn main() {
-    let crd = DeploymentRecord::crd();
+    let crd = ClassRuntime::crd();
     let yaml = serde_yaml::to_string(&crd).expect("serialize CRD to YAML");
     println!("{}", yaml);
 }
