+++ control-plane/oprc-crm/src/controller/cache.rs
@@ -2,11 +2,11 @@ use std::collections::HashMap;
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
-use crate::crd::deployment_record::DeploymentRecord;
+use crate::crd::class_runtime::ClassRuntime;
 
 #[derive(Clone)]
 pub struct DeploymentRecordCache(
-    pub Arc<RwLock<HashMap<String, DeploymentRecord>>>,
+    pub Arc<RwLock<HashMap<String, ClassRuntime>>>,
 );
 
 impl DeploymentRecordCache {
@@ -14,7 +14,7 @@ impl DeploymentRecordCache {
         Self(Arc::new(RwLock::new(HashMap::new())))
     }
 
-    pub async fn upsert(&self, key: String, dr: DeploymentRecord) {
+    pub async fn upsert(&self, key: String, dr: ClassRuntime) {
         let mut w = self.0.write().await;
         w.insert(key, dr);
     }
@@ -24,7 +24,7 @@ impl DeploymentRecordCache {
         w.remove(key);
     }
 
-    pub async fn list(&self) -> Vec<DeploymentRecord> {
+    pub async fn list(&self) -> Vec<ClassRuntime> {
         let r = self.0.read().await;
         r.values().cloned().collect()
     }
