+++ control-plane/oprc-crm/tests/it_kind_multi.rs
@@ -9,8 +9,9 @@ use kube::{
     Client,
     api::{Api, ListParams, PostParams},
 };
-use oprc_crm::crd::deployment_record::{
-    DeploymentRecord, DeploymentRecordSpec, FunctionSpec,
+use oprc_crm::crd::class_runtime::{
+    ClassRuntime as DeploymentRecord, ClassRuntimeSpec as DeploymentRecordSpec,
+    FunctionSpec,
 };
 use tracing::debug;
