+++ control-plane/oprc-crm/tests/it_multiple_functions.rs
@@ -5,8 +5,9 @@ use kube::{
     Client,
     api::{Api, ListParams, PostParams},
 };
-use oprc_crm::crd::deployment_record::{
-    DeploymentRecord, DeploymentRecordSpec, FunctionSpec,
+use oprc_crm::crd::class_runtime::{
+    ClassRuntime as DeploymentRecord, ClassRuntimeSpec as DeploymentRecordSpec,
+    FunctionSpec,
 };
 // std::time::Duration not used in this test
 mod common;
