+++ tools/oprc-cli/src/commands/mod.rs
@@ -1,17 +1,17 @@
 mod class;
+mod class_runtimes;
 mod clusters;
 mod context;
 mod deploy;
-mod deployment_records;
 mod deployment_status;
 mod function;
 mod package;
 
 pub use class::*;
+pub use class_runtimes::*;
 pub use clusters::*;
 pub use context::*;
 pub use deploy::*;
-pub use deployment_records::*;
 pub use deployment_status::*;
 pub use function::*;
 pub use package::*;
